TRECISION: Move MessageQueues into Scheduler
