+++ engines/trecision/scheduler.h
@@ -25,8 +25,48 @@
 
 #include "common/scummsys.h"
 
+#define MAXMESSAGE 128
+
 namespace Trecision {
+
 class TrecisionEngine;
+
+struct Message {
+	uint8 _class;    // message class
+	uint8 _event;    // message name
+	uint8 _priority; // message priority
+
+	uint8 _u8Param;
+	uint16 _u16Param1; // byte parameter 1
+	uint16 _u16Param2; // byte parameter 2
+	uint32 _u32Param;  // int parameter
+	uint32 _timestamp;
+
+	void set(Message *src) {
+		_class = src->_class;
+		_event = src->_event;
+		_priority = src->_priority;
+		_u8Param = src->_u8Param;
+		_u16Param1 = src->_u16Param1;
+		_u16Param2 = src->_u16Param2;
+		_u32Param = src->_u32Param;
+		_timestamp = src->_timestamp;
+	}
+};
+
+struct MessageQueue {
+	uint8 _head, _tail, _len;
+	Message *_event[MAXMESSAGE];
+
+	void orderEvents();
+	bool testEmptyQueues(uint8 cls);
+	bool testEmptyCharacterQueue4Script();
+	bool getMessage(Message **msg);
+	void initQueue();
+
+private:
+	uint8 predEvent(uint8 i);
+};
 
 class Scheduler {
 private:
@@ -35,6 +75,13 @@ private:
 	int8 _maxMessageCharacter;
 	uint8 _token;
 	uint8 _counter;
+
+	// Message system
+	Message _gameMsg[MAXMESSAGE];
+	Message _characterMsg[MAXMESSAGE];
+	Message _idleMsg;
+	MessageQueue _gameQueue;
+	MessageQueue _characterQueue;
 
 public:
 	Scheduler(TrecisionEngine *vm);
@@ -46,6 +93,11 @@ public:
 	void rightClick(uint16 x, uint16 y);
 	void mouseExamine(uint16 object);
 	void mouseOperate(uint16 object);
+
+	void init();
+	void resetQueues();
+	void initCharacterQueue();
+	bool testEmptyQueues();
 };
 
 } // End of namespace Trecision
