+++ engines/trecision/struct.h
@@ -175,43 +175,6 @@ struct SDText {
 	uint16 calcHeight(TrecisionEngine *vm);
 };
 
-struct Message {
-	uint8 _class;    // message class
-	uint8 _event;    // message name
-	uint8 _priority; // message priority
-
-	uint8 _u8Param;
-	uint16 _u16Param1; // byte parameter 1
-	uint16 _u16Param2; // byte parameter 2
-	uint32 _u32Param;  // int parameter
-	uint32 _timestamp;
-
-	void set(Message *src) {
-		_class = src->_class;
-		_event = src->_event;
-		_priority = src->_priority;
-		_u8Param = src->_u8Param;
-		_u16Param1 = src->_u16Param1;
-		_u16Param2 = src->_u16Param2;
-		_u32Param = src->_u32Param;
-		_timestamp = src->_timestamp;
-	}
-};
-
-struct MessageQueue {
-	uint8 _head, _tail, _len;
-	Message *_event[MAXMESSAGE];
-
-	void orderEvents();
-	bool testEmptyQueue(uint8 cls);
-	bool testEmptyCharacterQueue4Script();
-	bool getMessage(Message **msg);
-	void initQueue();
-
-private:
-	uint8 predEvent(uint8 i);
-};
-
 struct STexture {
 	int16 _dx, _dy, _angle;
 	uint8 *_texture;
